requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-core==1.32.0
blinker==1.9.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.10
dataclasses-json==0.6.7
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl#sha256=fec69fec52b1780f2d269d5af7582a5e28028738bd3190532459aeb473bfa3e7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
es_core_news_md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.8.0/es_core_news_md-3.8.0-py3-none-any.whl#sha256=478b8bb3f3e8eb149192f7d80e7cd64f990b7e9bccbc5329df41a57e86326be2
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain>=0.4.0
blis>=0.7.0
numpy>=1.22.4,<3.0.0
streamlit>=1.25.0
langchain-community==0.3.9
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langcodes==3.5.0
langsmith==0.1.146
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
narwhals==1.14.2
openai==1.56.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.50
pillow==11.0.0
preshed==3.0.9
presidio_analyzer==2.2.355
presidio_anonymizer==2.2.355
propcache==0.2.0
protobuf==5.28.3
pyarrow==18.1.0
pycryptodome==3.21.0
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
starlette==0.41.3
tenacity==9.0.0
thinc==8.3.2
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.0
weasel==0.4.1
websockets==14.1
wrapt==1.17.0
yarl==1.18.0
